Moved component out of generic components folder into it's screen folder
