+++ src/client/src/loading/hourglass-spinner.js
